hotfix: allow pushing via aliases using the old push syntax
