+++ packages/cli/src/__tests__/commands/push.test.ts
@@ -1,5 +1,5 @@
 import { Config } from '@redocly/openapi-core';
-import { getApiEntrypoint, getDestinationProps, handlePush } from '../../commands/push';
+import { getApiEntrypoint, getDestinationProps, handlePush, transformPush } from '../../commands/push';
 
 jest.mock('fs');
 jest.mock('node-fetch');
@@ -32,6 +32,72 @@ describe('push', () => {
       rootFilePath: 'filePath',
       version: '1.0.0',
     });
+  });
+});
+
+describe('transformPush', () => {
+  it('should adapt the existing syntax', () => {
+    const cb = jest.fn();
+    transformPush(cb)({
+      maybeEntrypointOrAliasOrDestination: 'openapi.yaml',
+      maybeDestination: '@testing_org/main@v1',
+    });
+    expect(cb).toBeCalledWith({
+      entrypoint: 'openapi.yaml',
+      destination: '@testing_org/main@v1',
+    });
+  });
+  it('should adapt the existing syntax (including branchName)', () => {
+    const cb = jest.fn();
+    transformPush(cb)({
+      maybeEntrypointOrAliasOrDestination: 'openapi.yaml',
+      maybeDestination: '@testing_org/main@v1',
+      maybeBranchName: 'other',
+    });
+    expect(cb).toBeCalledWith({
+      entrypoint: 'openapi.yaml',
+      destination: '@testing_org/main@v1',
+      branchName: 'other',
+    });
+  });
+  it('should use --branch option firstly', () => {
+    const cb = jest.fn();
+    transformPush(cb)({
+      maybeEntrypointOrAliasOrDestination: 'openapi.yaml',
+      maybeDestination: '@testing_org/main@v1',
+      maybeBranchName: 'other',
+      branch: 'priority-branch',
+    });
+    expect(cb).toBeCalledWith({
+      entrypoint: 'openapi.yaml',
+      destination: '@testing_org/main@v1',
+      branchName: 'priority-branch',
+    });
+  });
+  it('should work for a destination only', () => {
+    const cb = jest.fn();
+    transformPush(cb)({
+      maybeEntrypointOrAliasOrDestination: '@testing_org/main@v1',
+    });
+    expect(cb).toBeCalledWith({
+      destination: '@testing_org/main@v1',
+    });
+  });
+  it('should accept aliases for the old syntax', () => {
+    const cb = jest.fn();
+    transformPush(cb)({
+      maybeEntrypointOrAliasOrDestination: 'alias',
+      maybeDestination: '@testing_org/main@v1',
+    });
+    expect(cb).toBeCalledWith({
+      destination: '@testing_org/main@v1',
+      entrypoint: 'alias',
+    });
+  });
+  it('should accept no arguments at all', () => {
+    const cb = jest.fn();
+    transformPush(cb)({});
+    expect(cb).toBeCalledWith({});
   });
 });
